This function calls all of the f_val_calc.* files

SECTION Licor data assign variables

f_val_calc_all_driver(val_TDL, val_Licor, val_const, sw)

Arguments

val_TDL
xxxPARAMxxx

val_Licor
xxxPARAMxxx

val_const
xxxPARAMxxx

sw
xxxPARAMxxx

Value

val_temp xxxRETURNxxx

Details

SECTION Concentrations.

Calibration tank concentrations

Calibration tank gain. f_val_calc_gain

Calibration tank offset. f_val_calc_offset

Corrected reference values. f_val_calc_corrected

Corrected chamber values. f_val_calc_corrected

Total reference and chamber values. f_val_calc_total_mol_fraction_CO2

xi. f_val_calc_xi

SECTION A Photosynthesis

A, TDL photosynthesis. f_val_calc_TDL_A_photosynthesis

12A, TDL photosynthesis. f_val_calc_TDL_A_photosynthesis

13A, TDL photosynthesis. f_val_calc_TDL_A_photosynthesis

Total A LI6400 Photo LI6400 header, these values listed because I will likely use them in a plot

Delta from ratios in and out?, (Re/Ro)-1, is this really the same? f_val_calc_Delta_from_ratios_in_out

D from A ratio, (Ro/(13A/12A))-1, should be the same as Dobs above. f_val_calc_Delta_from_A_ratio

Select TDL or Licor based on switch

SECTION Delta

delta reference and chamber values. f_val_calc_delta_proportion

delta diff

Delta discrim

Dobs observed discrimination. f_val_calc_Delta_obs

Dobs per mil 1000*Dobs. f_val_calc_Delta_obs_permil

delta13C Assimilated, isotopic composition of assimilated sugars. f_val_calc_delta_13C_Assim

p (Co - Ce) / Co. f_val_calc_p

delta13C Respired, isotopic composition of respired CO2. f_val_calc_delta_13C_Resp

SECTION g conductance

gbw BLcond boundary layer conductance for water, Blcond is LI 6400 header

gbc BLcond/1.37 boundary layer conductance for CO2

gsw cond stomatal conductance for water, cond is LI6400 header

gsc gsw/1.6 stomatal conductance for CO2

gtc, total (stomatal and boundary layer) conductance for CO2

SECTION Cx and px (concentrations and pressures)

Ca = Co, ppm CO2 concentration above the leaf = concentration leaving the leaf chamber = ambient CO2 concentration

Cs, ppm CO2 concentration at the leaf surface, cs calculated from eq 40 Ball 1987 Ch 20 Stomatal Function, eds Zeiger, Farquhar, Cowan. f_val_calc_Cs

pa, partial pressure of CO2 above the leaf, Press is the atmospheric pressure value from the LI6400. f_val_calc_pp

ps, partial pressure of CO2 at the leaf surface, Press is the atmospheric pressure value from the LI6400. f_val_calc_pp NB: same formula as pa, but using Cs instead of Co

Ci, ppm CO2 concentration in the sub-stomatal cavities, ci calculated from eq 35 Ball 1987 Ch 20 Stomatal Function, eds Zeiger, Farquhar, Cowan. f_val_calc_Cs NB: same formula as Cs, but using gtc instead of gbc

pi, partial pressure of CO2 in the substomatal cavities, Press is the atmospheric pressure value from the LI6400. f_val_calc_pp NB: same formula as pa, but using Ci instead of Co

pi/pa total pi/pa or total Ci/Ca ratio of substomatal CO2 partial pressure to CO2 partial pressure above leaf = Ci/Ca ratio of mol fractions

SECTION Delta_i predictions

Delta_i simple for gm, predicted discrimination including boundary layer effects but not decarboxylation effects. f_val_calc_Delta_i_simple_for_gm

Delta_i simple for modeling a + (b-a) pi/pa predicted discrimination including boundary layer effects but using b adjustments to approximate effects of gm and decarboxylations. f_val_calc_Delta_i_simple_for_modeling

Delta_i complex for gm CHECK DERIVATION predicted discrimination including boundary layer effects AND decarboxylation effects. f_val_calc_Delta_i_complex_for_gm

Delta_i simple for gm-Dobs (Di simple for gm)-(Dobs per mil)

Delta_i complex for gm-Dobs (Di complex for gm)-(Dobs per mil)

SECTION gm mesophyll conductance

gm point simple, internal leaf (mesophyll) conductance calculated for every D value ignoring decarboxylation effects. f_val_calc_gm_point_simple

gm point complex, internal leaf (mesophyll) conductance calculated for every D value estimating decarboxylation effects. f_val_calc_gm_point_complex

select one of the calculated gm values to use

SECTION pc

pc using gm, total partial pressure of CO2 at the site of carboxylation, Press is the atmospheric pressure value from the LI6400. f_val_calc_pc_using_gm

pc using simple D for gm, includes boundary layer. f_val_calc_pc_using_simple_Delta_for_gm

pc using simple D for modeling. f_val_calc_pc_using_simple_Delta_for_modeling

pc using complex D, no decarboxylation [ab(pa-ps)+a(ps-pi)+pi(bs+al)-Dpa]/(bs+al-b) is this different from two up? They give different values but both may not be derived properly. f_val_calc_pc_using_complex_Delta_no_decarboxylation

pc using complex D, full model [ab(pa-ps)+a(ps-pi)+pi(bs+al)-(eRd/k+fG*)-Dpa]/(bs+al-b) includes boundary layer and decarboxylation effects. f_val_calc_pc_using_complex_Delta_full_model

select one of the calculated pc values to use

Cc (pc*10^6)/(Press*1000) ppm CO2 concentration at the site of carboxylation, generally meaning inside the chloroplast and ignoring PEPC in cytosol. f_val_calc_pp NB: same formula as pa, but using pc instead of Co