Reads TDL and Licor files, aligns them, calculates quantities of interest with bootstrap intervals.

Edit isogasex_templateX.xls and input parameters and TDL/Licor filenames. Run isogasex. See output in ./out directory.

isogasex(input_fn, path = getwd())

Arguments

input_fn
xxxPARAMxxx

path
xxxPARAMxxx

Value

NULL xxxRETURNxxx

Details

SECTION Preamble Set version. Display header.

Save original working directory, change to data dir. (This is a little hokey.) Change to data dir, then to out dir, then back to original at end.

Create output directory.

SECTION Read data.

Input Excel workbook. get_data

Assign variables. assign_variables

Create output directory (prefix) and update the process_info.txt.

Check that template version matches version of isogasex.

Set random seed for samples.

Read TDL and Licor data. read_TDL and read_Licor

SECTION TDL Plot and interpolate.

BEGIN If using TDL:

_ Print lines with missing values in TDL file and fix them. fix_missing_TDL_values

_ Extract cycle timing, write to file. extract_cycle_timing

_ Determine the last index for admissible values to calculate means. last_TDL_index_for_means

_ Plot each cycle of TDL measurements (visual diagnostics). plot_data_cycles

_ Calculate the mean and variance for the TDL data based on last measurements. calc_mean_TDL

_ Interpolate TDL tank and reference values. interp_TDL_tanks_ref

_ Summary values for TDL interpolated tanks and reference. calc_mean_TDL_interp_tanks_ref

_ Calculate mean Variance for interp values for Par BS of tanks and reference.

END

SECTION Align TDL and Licor

Determine overlapping time window of TDL and Licor measurements. align_TDL_Licor_times

Reduce data to overlapping time window and interp Licor values to TDL timepoints. time_window_TDL_Licor_interp

Plot each cycle of TDL measurements (visual diagnostics). plot_data_cycles

Calculate the mean and variance for the Licor data based on last measurements. calc_mean_Licor

SECTION Mean Calculations

Assign TDL and Licor values to val variable names, raw and summarized values. val_TDL_Licor_variables

summary values (mean). f_val_calc_all_driver

All time point values. f_val_calc_all_driver

SECTION Write out results

Summary files

TDL file. write_summary_TDL_file

Licor file. write_summary_Licor_file

Summary Calculation file. write_summary_Calc_file

All Calculation file. write_all_Calc_file

plot all the calculated values. plot_val_calc_sum

SECTION Bootstrap

BEGIN Bootstrap Calculations

_ Generate NP and Par BS samples for observed values.

. f_init_bs_matrix

. f_bs_iter_TDL_Licor

. f_bs_save_TDL_Licor

. f_val_calc_all_driver

. f_val_bs_matrix

_ calculate central 95% intervals.

_ create CI for each calculated value. f_val_bs_CI

_ plot all variables with bs values, mean value, and CI intervals. f_plot_CI_individuals_driver

_ write out central 95% interval, SD,

_ Summary files. write_CI_TDL_file

_ TDL file

_ Licor file. write_CI_Licor_file

_ Calculation file. write_CI_Calc_file

END

Save workspace.

SECTION Wrap it up. Move selected plot files into subdirectories. move_plot_files

Copy template file to out dir.

Change back to original dir.

Complete.