This function plots all the calculated values

most plots are not created when not using Licor file

plot_val_calc_sum(val_TDL, val_Licor, val_const, val_temp,
  plot_format_list, output_fn_prefix, sw)

Arguments

val_TDL
xxxPARAMxxx

val_Licor
xxxPARAMxxx

val_const
xxxPARAMxxx

val_temp
xxxPARAMxxx

plot_format_list
xxxPARAMxxx

output_fn_prefix
xxxPARAMxxx

sw
xxxPARAMxxx

Value

NULL xxxRETURNxxx

Details

Plots many variables by calling the plot.* functions.

Plot active gains, offsets, and corrected magnitudes. plot_gain_offset

Plot corrected 12CO2 and 13CO2, total, and difference for reference and sample. plot_corrected_total_diff_xi

Plot A photosynthesis. plot_A_photosynthesis

Plot Licor Temp Light values. plot_Licor_temp_light

Plot Licor uin, xin, La, Atm_press values 7/21/2012. plot_Licor_flow_press

Plot delta, Delta, and p. plot_delta_Delta_p

Plot Ca Cs, CO2 concentrations above the leaf and at the leaf surface. plot_Ca_Cs

Plot Pa Ps, partial pressure of CO2 above the leaf and at the leaf surface. plot_pa_ps

Plot gbc, gsc, gtc: boundary layer, stomatal, and total conductance for CO2. plot_gbc_gsc_gtc

Plot Ci, Pi, CO2 concentration and partial pressure of CO2 in the substomatal cavities. plot_Ci_pi

Plot Totalpi_pa, ratio of substomatal CO2 partial pressure to CO2 partial pressure above leaf = Ci/Ca ratio of mol fractions. plot_pi_pa

Plot Delta_i, predicted discrimination. plot_Delta_i

Plot gm point simple and complex, internal leaf (mesophyll) conductance calculated for every D value ignoring and estimating decarboxylation effects. plot_gm

Plot pc, total partial pressure of CO2 at the site of carboxylation. plot_pc_total

Plot pc, simple and complex. plot_pc_simple_complex

Plot Cc, ppm CO2 concentration at the site of carboxylation, generally meaning inside the chloroplast and ignoring PEPC in cytosol. plot_Cc_total